+++ lib/src/functions/testing/utils.dart
@@ -6,6 +6,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+/// [debugPrint] rename
+/// 100% so I can search for [debugPrint] in my code to find the temporary ones
+void ezLog(String message, {int? wrapWidth}) =>
+    debugPrint(message, wrapWidth: wrapWidth);
+
 /// For integration tests
 /// Wait for a desired number of [seconds]
 Future<void> pause(int seconds) =>